requirements.txt
filetype==1.2.0
pillow==11.0.0
